requirements.txt
streamlit
groq
python-dotenv
pypdf
pandas
openpyxl
python-docx